requirements.txt
PyYAML==3.11
pep8==1.5.7
pylint==1.2.1
pytest==2.5.2
pytest-cov==1.7.0
tox==1.7.2
twine==1.3.1
wheel==0.24.0
